refactor(StageService): 修改StageSetResolution的分辨率参数类型和写入格式
将分辨率参数从float32改为uint32，同时调整PLC写入时的数据格式为UInt32，匹配硬件寄存器的实际数据类型要求

+++ services/components/StageService.go
@@ -241,14 +241,14 @@ func (this *StageService) StageSetSpeed(Axis string, speed float32) error {
 }
 
 // StageSetResolution 设置指定轴的分辨率
-func (this *StageService) StageSetResolution(Axis string, resolution float32) error {
+func (this *StageService) StageSetResolution(Axis string, resolution uint32) error {
 	if debug {
 		fmt.Println("StageSetResolution", Axis, resolution)
 	}
 	if _, ok := this.motors[Axis]; !ok {
 		return errors.New("unknown axis: " + Axis)
 	}
-	return this.plc.WriteRegister(XinJie.HD(this.motors[Axis].ADDR.RESOLUTION_HD), resolution, XinJie.Float32, false)
+	return this.plc.WriteRegister(XinJie.HD(this.motors[Axis].ADDR.RESOLUTION_HD), resolution, XinJie.UInt32, false)
 }
 
 func (this *StageService) SetAlarmLED(alarm bool) error {
